chore: Move extension files to repo root
I used Yeoman to create the extension. As you can see in the initial commit 9a697eba97c201862d6edef7ec158e6a197863dc that involved installing the generators `generator-code` package.

But after publishing this package I can see that the images don't work in the README. The assumption and common practice seems to be to have the extension in the repo root.